add echo statements for current executing line

+++ setup.sh
@@ -6,8 +6,10 @@ DOTFILES_DIR=`pwd`
 ## BREW
 
 # Install command line tools
+echo "Checking for xcode command line tools ..."
 xcode-select -p 1>/dev/null
 if [[ $? != 0 ]] ; then
+    echo "Installing xcode command line tools ..."
     # Install command line tools
     xcode-select --install
     # Accept command line tools license
@@ -21,9 +23,11 @@ which -s brew
 
 if [[ $? != 0 ]] ; then
     # Install Homebrew
+    echo "Installing brew ..."
     /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
 else
     # Update brew
+    echo "Updating brew ..."
     brew update
 fi
 
@@ -31,15 +35,20 @@ fi
 ## ZSH
 
 # Install pure prompt
+echo "Installing pure ..."
 brew install pure
 
 # Link zshrc
 if [ -f ~/.zshrc ]; then
     # this is in case these is an existing useful zshrc which we want to extend
     # example: on a work laptop
-    ln -s $DOTFILES_DIR/zshrc.symlink ~/.custom-zshrc.sh
-    echo "source $HOME/.custom-zshrc.sh" >> ~/.zshrc
+    echo "Linking custom zshrc ..."
+    if ! [ -L ~/.custom-zshrc.sh ]; then
+        ln -s $DOTFILES_DIR/zshrc.symlink ~/.custom-zshrc.sh
+        echo "source $HOME/.custom-zshrc.sh" >> ~/.zshrc
+    fi
 else
+    echo "Linking zshrc ..."
     ln -s $DOTFILES_DIR/zshrc.symlink ~/.zshrc
 fi
 
@@ -47,15 +56,18 @@ fi
 ## PYTHON
 
 # Install python
+echo "Installing python ..."
 brew install pyenv
 brew install zlib
 pyenv install 3.10.4
 pyenv global 3.10.4
 
 # Install poetry
-curl -sSL https://raw.githubusercontent.com/python-poetry/poetry/master/get-poetry.py | python -
+echo "Installing poetry ..."
+curl -sSL https://install.python-poetry.org | python3 -
 
 # Install python tools
+echo "Installing python tools ..."
 pip install --upgrade pip
 pip install ipython
 
@@ -63,6 +75,7 @@ pip install ipython
 ## LUA
 
 # Install lua
+echo "Setting up lua ..."
 brew install lua
 brew install luarocks
 brew install lua-language-server
@@ -71,8 +84,10 @@ brew install lua-language-server
 ## NEOVIM
 
 # Install neovim stuff
+echo "Installing neovim and utilities..."
 brew install neovim
 brew install fzf
+echo "Check if fzf keybindings need to be installed!"
 # run /usr/local/opt/fzf/install to setup keybindings
 brew install ripgrep
 brew install the_silver_searcher
@@ -89,6 +104,7 @@ sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.
     #https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 
 # Link neovim config
+echo "linking neovim config ..."
 mkdir -p ~/.config/nvim/lua
 
 # first backup any existing config
@@ -96,15 +112,15 @@ if [ -f ~/.config/nvim/init.vim ]; then
     mv ~/.config/nvim/init.vim ~/.config/nvim/init.vim.backup
 fi
 
-if [ -f ~/.config/nvim/lua/config.lua]; then
+if [ -f ~/.config/nvim/lua/config.lua ]; then
     mv ~/.config/nvim/lua/config.lua ~/.config/nvim/lua/config.lua.backup
 fi
 
-if [ -f ~/.config/nvim/lua/lsp_config.lua]; then
+if [ -f ~/.config/nvim/lua/lsp_config.lua ]; then
     mv ~/.config/nvim/lua/lsp_config.lua ~/.config/nvim/lua/lsp_config.lua.backup
 fi
 
-if [ -f ~/.config/nvim/lua/cmp_config.lua]; then
+if [ -f ~/.config/nvim/lua/cmp_config.lua ]; then
     mv ~/.config/nvim/lua/cmp_config.lua ~/.config/nvim/lua/cmp_config.lua.backup
 fi
 
@@ -115,13 +131,14 @@ ln -s $DOTFILES_DIR/nvim/lua/lsp_config.lua ~/.config/nvim/lua/lsp_config.lua
 ln -s $DOTFILES_DIR/nvim/lua/cmp_config.lua ~/.config/nvim/lua/cmp_config.lua
 
 # Install nvim plugins
+echo "Installing neovim plugins ..."
 nvim +PlugInstall +qall
-#mvim +PlugInstall +qall
 
 
 ## GOLANG
 
 # Install go
+echo "Installing golang ..."
 mkdir -p $HOME/go/{bin,src}
 brew install golang
 
@@ -129,6 +146,7 @@ brew install golang
 ## GIT
 
 # Link gitconfig
+echo "Setup git ..."
 if [ -f ~/.gitconfig ]; then
     mv ~/.gitconfig ~/.gitconfig.backup
 fi
@@ -138,6 +156,7 @@ ln -s $DOTFILES_DIR/gitconfig.symlink ~/.gitconfig
 ## TERMINAL
 
 # Install fira code
+echo "Installing fira code font ..."
 brew tap homebrew/cask-fonts
 brew install --cask font-fira-code-nerd-font
 
@@ -145,6 +164,7 @@ brew install --cask font-fira-code-nerd-font
 ## UTILITIES
 
 # Install utilities
+echo "Installing utilities ..."
 brew install wget
 brew install htop
